+++ CHANGES.md
@@ -3,6 +3,7 @@
 
 ## Bugfixes
 - `StructureUtil.getPathOfFieldContaining()` and therefore `@Role` fields ignored fields where the value was actually a subclass of the field type.
+- CopyFrom and templating failed on primitive values. See [#400](https://github.com/klum-dsl/klum-ast/issues/400)
 
 # 2.1.4/2.1.5
 - Prevent ConcurrentModificationExceptions when calling `applyLater` from a lifecycle method